Fix some size_t implicitly converted to UniValue
The compiler could complain about an ambiguous implicit conversion.

+++ src/rpc/blockchain.cpp
@@ -2934,7 +2934,7 @@ static UniValue getoooblocks(const JSONRPCRequest& request)
         fVerbose = request.params[0].get_bool();
 
     if (!fVerbose)
-        return CountOoOBlocks();
+        return (uint64_t)CountOoOBlocks();
     else {
         const auto ooob_map = GetOoOBlockMap();
         UniValue ret(UniValue::VOBJ);
